Split Model to BaseModel and Model
Apply suggestions from code review
