+++ system/Model.php
@@ -20,15 +20,9 @@ use CodeIgniter\Database\ConnectionInterface;
 use CodeIgniter\Database\Exceptions\DatabaseException;
 use CodeIgniter\Database\Exceptions\DataException;
 use CodeIgniter\Exceptions\ModelException;
-use CodeIgniter\I18n\Time;
-use CodeIgniter\Pager\Pager;
 use CodeIgniter\Validation\ValidationInterface;
 use Config\Database;
-use Config\Services;
-use ReflectionClass;
 use ReflectionException;
-use ReflectionProperty;
-use stdClass;
 
 /**
  * Class Model
@@ -47,23 +41,8 @@ use stdClass;
  *
  * @mixin BaseBuilder
  */
-class Model
+class Model extends BaseModel
 {
-	/**
-	 * Pager instance.
-	 * Populated after calling $this->paginate()
-	 *
-	 * @var Pager
-	 */
-	public $pager;
-
-	/**
-	 * Name of database table
-	 *
-	 * @var string
-	 */
-	protected $table;
-
 	/**
 	 * The table's primary key.
 	 *
@@ -79,54 +58,6 @@ class Model
 	protected $useAutoIncrement = true;
 
 	/**
-	 * Last insert ID
-	 *
-	 * @var integer|string
-	 */
-	protected $insertID = 0;
-
-	/**
-	 * The Database connection group that
-	 * should be instantiated.
-	 *
-	 * @var string
-	 */
-	protected $DBGroup;
-
-	/**
-	 * The format that the results should be returned as.
-	 * Will be overridden if the as* methods are used.
-	 *
-	 * @var string
-	 */
-	protected $returnType = 'array';
-
-	/**
-	 * If this model should use "softDeletes" and
-	 * simply set a date when rows are deleted, or
-	 * do hard deletes.
-	 *
-	 * @var boolean
-	 */
-	protected $useSoftDeletes = false;
-
-	/**
-	 * An array of field names that are allowed
-	 * to be set by the user in inserts/updates.
-	 *
-	 * @var array
-	 */
-	protected $allowedFields = [];
-
-	/**
-	 * If true, will set created_at, and updated_at
-	 * values during insert and update routines.
-	 *
-	 * @var boolean
-	 */
-	protected $useTimestamps = false;
-
-	/**
 	 * The type of column that created_at and updated_at
 	 * are expected to.
 	 *
@@ -135,51 +66,6 @@ class Model
 	 * @var string
 	 */
 	protected $dateFormat = 'datetime';
-
-	/**
-	 * The column used for insert timestamps
-	 *
-	 * @var string
-	 */
-	protected $createdField = 'created_at';
-
-	/**
-	 * The column used for update timestamps
-	 *
-	 * @var string
-	 */
-	protected $updatedField = 'updated_at';
-
-	/**
-	 * Used by withDeleted to override the
-	 * model's softDelete setting.
-	 *
-	 * @var boolean
-	 */
-	protected $tempUseSoftDeletes;
-
-	/**
-	 * The column used to save soft delete state
-	 *
-	 * @var string
-	 */
-	protected $deletedField = 'deleted_at';
-
-	/**
-	 * Used by asArray and asObject to provide
-	 * temporary overrides of model default.
-	 *
-	 * @var string
-	 */
-	protected $tempReturnType;
-
-	/**
-	 * Whether we should limit fields in inserts
-	 * and updates to those available in $allowedFields or not.
-	 *
-	 * @var boolean
-	 */
-	protected $protectFields = true;
 
 	/**
 	 * Database Connection
@@ -196,153 +82,16 @@ class Model
 	protected $builder;
 
 	/**
-	 * Rules used to validate data in insert, update, and save methods.
-	 * The array must match the format of data passed to the Validation
-	 * library.
-	 *
-	 * @var array|string
-	 */
-	protected $validationRules = [];
-
-	/**
-	 * Contains any custom error messages to be
-	 * used during data validation.
-	 *
-	 * @var array
-	 */
-	protected $validationMessages = [];
-
-	/**
-	 * Skip the model's validation. Used in conjunction with skipValidation()
-	 * to skip data validation for any future calls.
-	 *
-	 * @var boolean
-	 */
-	protected $skipValidation = false;
-
-	/**
-	 * Whether rules should be removed that do not exist
-	 * in the passed in data. Used between inserts/updates.
-	 *
-	 * @var boolean
-	 */
-	protected $cleanValidationRules = true;
-
-	/**
-	 * Our validator instance.
-	 *
-	 * @var ValidationInterface
-	 */
-	protected $validation;
-
-	/*
-	 * Callbacks.
-	 *
-	 * Each array should contain the method names (within the model)
-	 * that should be called when those events are triggered.
-	 *
-	 * "Update" and "delete" methods are passed the same items that
-	 * are given to their respective method.
-	 *
-	 * "Find" methods receive the ID searched for (if present), and
-	 * 'afterFind' additionally receives the results that were found.
-	 */
-
-	/**
-	 * Whether to trigger the defined callbacks
-	 *
-	 * @var boolean
-	 */
-	protected $allowCallbacks = true;
-
-	/**
-	 * Used by allowCallbacks() to override the
-	 * model's allowCallbacks setting.
-	 *
-	 * @var boolean
-	 */
-	protected $tempAllowCallbacks;
-
-	/**
-	 * Callbacks for beforeInsert
-	 *
-	 * @var array
-	 */
-	protected $beforeInsert = [];
-
-	/**
-	 * Callbacks for afterInsert
-	 *
-	 * @var array
-	 */
-	protected $afterInsert = [];
-
-	/**
-	 * Callbacks for beforeUpdate
-	 *
-	 * @var array
-	 */
-	protected $beforeUpdate = [];
-
-	/**
-	 * Callbacks for afterUpdate
-	 *
-	 * @var array
-	 */
-	protected $afterUpdate = [];
-
-	/**
-	 * Callbacks for beforeFind
-	 *
-	 * @var array
-	 */
-	protected $beforeFind = [];
-
-	/**
-	 * Callbacks for afterFind
-	 *
-	 * @var array
-	 */
-	protected $afterFind = [];
-
-	/**
-	 * Callbacks for beforeDelete
-	 *
-	 * @var array
-	 */
-	protected $beforeDelete = [];
-
-	/**
-	 * Callbacks for afterDelete
-	 *
-	 * @var array
-	 */
-	protected $afterDelete = [];
-
-	/**
-	 * Holds information passed in via 'set'
-	 * so that we can capture it (not the builder)
-	 * and ensure it gets validated first.
-	 *
-	 * @var array
-	 */
-	protected $tempData = [];
-
-	/**
 	 * Model constructor.
 	 *
-	 * @param ConnectionInterface|null $db
-	 * @param ValidationInterface|null $validation
+	 * @param ConnectionInterface|null $db         DB Connection
+	 * @param ValidationInterface|null $validation Validation
 	 */
 	public function __construct(ConnectionInterface &$db = null, ValidationInterface $validation = null)
 	{
-		$db = $db ?? Database::connect($this->DBGroup);
+		parent::__construct($db, $validation);
 
-		$this->db                 = &$db;
-		$this->tempReturnType     = $this->returnType;
-		$this->tempUseSoftDeletes = $this->useSoftDeletes;
-		$this->tempAllowCallbacks = $this->allowCallbacks;
-		$this->validation         = $validation ?? Services::validation(null, false);
+		$this->db = $db ?? Database::connect($this->DBGroup);
 	}
 
 	//--------------------------------------------------------------------
@@ -351,31 +100,15 @@ class Model
 
 	/**
 	 * Fetches the row of database from $this->table with a primary key
-	 * matching $id.
+	 * matching $id. This methods works only with dbCalls
 	 *
-	 * @param array|integer|string|null $id One primary key or an array of primary keys
+	 * @param boolean                   $singleton Single or multiple results
+	 * @param array|integer|string|null $id        One primary key or an array of primary keys
 	 *
 	 * @return array|object|null    The resulting row of data, or null.
 	 */
-	public function find($id = null)
+	protected function _find(bool $singleton, $id = null)
 	{
-		$singleton = is_numeric($id) || is_string($id);
-
-		if ($this->tempAllowCallbacks)
-		{
-			// Call the before event and check for a return
-			$eventData = $this->trigger('beforeFind', [
-				'id'        => $id,
-				'method'    => 'find',
-				'singleton' => $singleton,
-			]);
-
-			if (! empty($eventData['returnData']))
-			{
-				return $eventData['data'];
-			}
-		}
-
 		$builder = $this->builder();
 
 		if ($this->tempUseSoftDeletes)
@@ -400,72 +133,32 @@ class Model
 			$row = $builder->get()->getResult($this->tempReturnType);
 		}
 
-		$eventData = [
-			'id'        => $id,
-			'data'      => $row,
-			'method'    => 'find',
-			'singleton' => $singleton,
-		];
-
-		if ($this->tempAllowCallbacks)
-		{
-			$eventData = $this->trigger('afterFind', $eventData);
-		}
-
-		$this->tempReturnType     = $this->returnType;
-		$this->tempUseSoftDeletes = $this->useSoftDeletes;
-		$this->tempAllowCallbacks = $this->allowCallbacks;
-
-		return $eventData['data'];
+		return $row;
 	}
 
 	/**
 	 * Fetches the column of database from $this->table
 	 *
-	 * @param string $columnName
+	 * @param string $columnName Column Name
 	 *
 	 * @return array|null   The resulting row of data, or null if no data found.
-	 * @throws DataException
 	 */
-	public function findColumn(string $columnName)
+	protected function _findColumn(string $columnName)
 	{
-		if (strpos($columnName, ',') !== false)
-		{
-			throw DataException::forFindColumnHaveMultipleColumns();
-		}
-
-		$resultSet = $this->select($columnName)->asArray()->find();
-
-		return $resultSet ? array_column($resultSet, $columnName) : null;
+		return $this->select($columnName)->asArray()->find();
 	}
 
 	/**
 	 * Works with the current Query Builder instance to return
 	 * all results, while optionally limiting them.
 	 *
-	 * @param integer $limit
-	 * @param integer $offset
+	 * @param integer $limit  Limit
+	 * @param integer $offset Offset
 	 *
 	 * @return array
 	 */
-	public function findAll(int $limit = 0, int $offset = 0)
+	protected function _findAll(int $limit = 0, int $offset = 0)
 	{
-		if ($this->tempAllowCallbacks)
-		{
-			// Call the before event and check for a return
-			$eventData = $this->trigger('beforeFind', [
-				'method'    => 'findAll',
-				'limit'     => $limit,
-				'offset'    => $offset,
-				'singleton' => false,
-			]);
-
-			if (! empty($eventData['returnData']))
-			{
-				return $eventData['data'];
-			}
-		}
-
 		$builder = $this->builder();
 
 		if ($this->tempUseSoftDeletes)
@@ -473,28 +166,9 @@ class Model
 			$builder->where($this->table . '.' . $this->deletedField, null);
 		}
 
-		$row = $builder->limit($limit, $offset)
-				->get()
-				->getResult($this->tempReturnType);
-
-		$eventData = [
-			'data'      => $row,
-			'limit'     => $limit,
-			'offset'    => $offset,
-			'method'    => 'findAll',
-			'singleton' => false,
-		];
-
-		if ($this->tempAllowCallbacks)
-		{
-			$eventData = $this->trigger('afterFind', $eventData);
-		}
-
-		$this->tempReturnType     = $this->returnType;
-		$this->tempUseSoftDeletes = $this->useSoftDeletes;
-		$this->tempAllowCallbacks = $this->allowCallbacks;
-
-		return $eventData['data'];
+		return $builder->limit($limit, $offset)
+			->get()
+			->getResult($this->tempReturnType);
 	}
 
 	/**
@@ -503,22 +177,8 @@ class Model
 	 *
 	 * @return array|object|null
 	 */
-	public function first()
+	protected function _first()
 	{
-		if ($this->tempAllowCallbacks)
-		{
-			// Call the before event and check for a return
-			$eventData = $this->trigger('beforeFind', [
-				'method'    => 'first',
-				'singleton' => true,
-			]);
-
-			if (! empty($eventData['returnData']))
-			{
-				return $eventData['data'];
-			}
-		}
-
 		$builder = $this->builder();
 
 		if ($this->tempUseSoftDeletes)
@@ -540,43 +200,7 @@ class Model
 			$builder->orderBy($this->table . '.' . $this->primaryKey, 'asc');
 		}
 
-		$eventData = [
-			'data'      => $builder->limit(1, 0)->get()->getFirstRow($this->tempReturnType),
-			'method'    => 'first',
-			'singleton' => true,
-		];
-
-		if ($this->tempAllowCallbacks)
-		{
-			$eventData = $this->trigger('afterFind', $eventData);
-		}
-
-		$this->tempReturnType     = $this->returnType;
-		$this->tempUseSoftDeletes = $this->useSoftDeletes;
-		$this->tempAllowCallbacks = $this->allowCallbacks;
-
-		return $eventData['data'];
-	}
-
-	/**
-	 * Captures the builder's set() method so that we can validate the
-	 * data here. This allows it to be used with any of the other
-	 * builder methods and still get validated data, like replace.
-	 *
-	 * @param mixed   $key    Field name, or an array of field/value pairs
-	 * @param string  $value  Field value, if $key is a single field
-	 * @param boolean $escape Whether to escape values and identifiers
-	 *
-	 * @return $this
-	 */
-	public function set($key, ?string $value = '', bool $escape = null)
-	{
-		$data = is_array($key) ? $key : [$key => $value];
-
-		$this->tempData['escape'] = $escape;
-		$this->tempData['data']   = array_merge($this->tempData['data'] ?? [], $data);
-
-		return $this;
+		return $builder->limit(1, 0)->get()->getFirstRow($this->tempReturnType);
 	}
 
 	/**
@@ -586,18 +210,12 @@ class Model
 	 * you must ensure that the class will provide access to the class
 	 * variables, even if through a magic method.
 	 *
-	 * @param array|object $data
+	 * @param array|object $data Data
 	 *
 	 * @return boolean
-	 * @throws ReflectionException
 	 */
-	public function save($data): bool
+	protected function _save($data): bool
 	{
-		if (empty($data))
-		{
-			return true;
-		}
-
 		// When useAutoIncrement feature is disabled check
 		// in the database if given record already exists
 		if (! $makeUpdate = $this->useAutoIncrement)
@@ -640,180 +258,30 @@ class Model
 				$response = true;
 			}
 		}
-
 		return $response;
-	}
-
-	/**
-	 * Takes a class an returns an array of it's public and protected
-	 * properties as an array suitable for use in creates and updates.
-	 *
-	 * @param string|object $data
-	 * @param string|null   $primaryKey
-	 * @param string        $dateFormat
-	 * @param boolean       $onlyChanged
-	 *
-	 * @return array
-	 * @throws ReflectionException
-	 */
-	public static function classToArray($data, $primaryKey = null, string $dateFormat = 'datetime', bool $onlyChanged = true): array
-	{
-		if (method_exists($data, 'toRawArray'))
-		{
-			$properties = $data->toRawArray($onlyChanged);
-
-			// Always grab the primary key otherwise updates will fail.
-			if (! empty($properties) && ! empty($primaryKey) && ! in_array($primaryKey, $properties, true) && ! empty($data->{$primaryKey}))
-			{
-				$properties[$primaryKey] = $data->{$primaryKey};
-			}
-		}
-		else
-		{
-			$mirror = new ReflectionClass($data);
-			$props  = $mirror->getProperties(ReflectionProperty::IS_PUBLIC | ReflectionProperty::IS_PROTECTED);
-
-			$properties = [];
-
-			// Loop over each property,
-			// saving the name/value in a new array we can return.
-			foreach ($props as $prop)
-			{
-				// Must make protected values accessible.
-				$prop->setAccessible(true);
-				$properties[$prop->getName()] = $prop->getValue($data);
-			}
-		}
-
-		// Convert any Time instances to appropriate $dateFormat
-		if ($properties)
-		{
-			foreach ($properties as $key => $value)
-			{
-				if ($value instanceof Time)
-				{
-					switch ($dateFormat)
-					{
-						case 'datetime':
-							$converted = $value->format('Y-m-d H:i:s');
-							break;
-						case 'date':
-							$converted = $value->format('Y-m-d');
-							break;
-						case 'int':
-							$converted = $value->getTimestamp();
-							break;
-						default:
-							$converted = (string) $value;
-					}
-
-					$properties[$key] = $converted;
-				}
-			}
-		}
-
-		return $properties;
-	}
-
-	/**
-	 * Returns last insert ID or 0.
-	 *
-	 * @return integer|string
-	 */
-	public function getInsertID()
-	{
-		return is_numeric($this->insertID) ? (int) $this->insertID : $this->insertID;
 	}
 
 	/**
 	 * Inserts data into the current table. If an object is provided,
 	 * it will attempt to convert it to an array.
 	 *
-	 * @param array|object $data
-	 * @param boolean      $returnID Whether insert ID should be returned or not.
+	 * @param array|object $data   Data
+	 * @param boolean|null $escape Escape
 	 *
 	 * @return BaseResult|integer|string|false
-	 * @throws ReflectionException
 	 */
-	public function insert($data = null, bool $returnID = true)
+	protected function _insert($data, bool $escape = null)
 	{
-		$escape = null;
-
-		$this->insertID = 0;
-
-		if (empty($data))
-		{
-			$data           = $this->tempData['data'] ?? null;
-			$escape         = $this->tempData['escape'] ?? null;
-			$this->tempData = [];
-		}
-
-		if (empty($data))
-		{
-			throw DataException::forEmptyDataset('insert');
-		}
-
-		// If $data is using a custom class with public or protected
-		// properties representing the table elements, we need to grab
-		// them as an array.
-		if (is_object($data) && ! $data instanceof stdClass)
-		{
-			$data = static::classToArray($data, $this->primaryKey, $this->dateFormat, false);
-		}
-
-		// If it's still a stdClass, go ahead and convert to
-		// an array so doProtectFields and other model methods
-		// don't have to do special checks.
-		if (is_object($data))
-		{
-			$data = (array) $data;
-		}
-
-		if (empty($data))
-		{
-			throw DataException::forEmptyDataset('insert');
-		}
-
-		// Validate data before saving.
-		if (! $this->skipValidation && ! $this->cleanRules()->validate($data))
-		{
-			return false;
-		}
-
-		// Must be called first so we don't
-		// strip out created_at values.
-		$data = $this->doProtectFields($data);
-
-		// Set created_at and updated_at with same time
-		$date = $this->setDate();
-
-		if ($this->useTimestamps && $this->createdField && ! array_key_exists($this->createdField, $data))
-		{
-			$data[$this->createdField] = $date;
-		}
-
-		if ($this->useTimestamps && $this->updatedField && ! array_key_exists($this->updatedField, $data))
-		{
-			$data[$this->updatedField] = $date;
-		}
-
-		$eventData = ['data' => $data];
-
-		if ($this->tempAllowCallbacks)
-		{
-			$eventData = $this->trigger('beforeInsert', $eventData);
-		}
-
 		// Require non empty primaryKey when
 		// not using auto-increment feature
-		if (! $this->useAutoIncrement && empty($eventData['data'][$this->primaryKey]))
+		if (! $this->useAutoIncrement && empty($data[$this->primaryKey]))
 		{
 			throw DataException::forEmptyPrimaryKey('insert');
 		}
 
 		// Must use the set() method to ensure objects get converted to arrays
 		$result = $this->builder()
-			->set($eventData['data'], '', $escape)
+			->set($data, '', $escape)
 			->insert();
 
 		// If insertion succeeded then save the insert ID
@@ -821,7 +289,7 @@ class Model
 		{
 			if (! $this->useAutoIncrement)
 			{
-				$this->insertID = $eventData['data'][$this->primaryKey];
+				$this->insertID = $data[$this->primaryKey];
 			}
 			else
 			{
@@ -829,90 +297,36 @@ class Model
 			}
 		}
 
-		$eventData = [
-			'id'     => $this->insertID,
-			'data'   => $eventData['data'],
-			'result' => $result,
-		];
-
-		if ($this->tempAllowCallbacks)
-		{
-			// Trigger afterInsert events with the inserted data and new ID
-			$this->trigger('afterInsert', $eventData);
-		}
-
-		$this->tempAllowCallbacks = $this->allowCallbacks;
-
-		// If insertion failed, get out of here
-		if (! $result)
-		{
-			return $result;
-		}
-
-		// otherwise return the insertID, if requested.
-		return $returnID ? $this->insertID : $result;
+		return $result;
 	}
 
 	/**
 	 * Compiles batch insert strings and runs the queries, validating each row prior.
 	 *
-	 * @param array   $set       An associative array of insert values
-	 * @param boolean $escape    Whether to escape values and identifiers
-	 * @param integer $batchSize The size of the batch to run
-	 * @param boolean $testing   True means only number of records is returned, false will execute the query
+	 * @param array|null $set       An associative array of insert values
+	 * @param boolean    $escape    Whether to escape values and identifiers
+	 * @param integer    $batchSize The size of the batch to run
+	 * @param boolean    $testing   True means only number of records is returned, false will execute the query
 	 *
 	 * @return integer|boolean Number of rows inserted or FALSE on failure
+	 * @throws ReflectionException ReflectionException.
 	 */
-	public function insertBatch(array $set = null, bool $escape = null, int $batchSize = 100, bool $testing = false)
+	protected function _insertBatch(
+		array $set = null,
+		bool $escape = null,
+		int $batchSize = 100,
+		bool $testing = false
+	)
 	{
 		if (is_array($set))
 		{
 			foreach ($set as &$row)
 			{
-				// If $data is using a custom class with public or protected
-				// properties representing the table elements, we need to grab
-				// them as an array.
-				if (is_object($row) && ! $row instanceof stdClass)
-				{
-					$row = static::classToArray($row, $this->primaryKey, $this->dateFormat, false);
-				}
-
-				// If it's still a stdClass, go ahead and convert to
-				// an array so doProtectFields and other model methods
-				// don't have to do special checks.
-				if (is_object($row))
-				{
-					$row = (array) $row;
-				}
-
-				// Validate every row..
-				if (! $this->skipValidation && ! $this->cleanRules()->validate($row))
-				{
-					return false;
-				}
-
-				// Must be called first so we don't
-				// strip out created_at values.
-				$row = $this->doProtectFields($row);
-
 				// Require non empty primaryKey when
 				// not using auto-increment feature
 				if (! $this->useAutoIncrement && empty($row[$this->primaryKey]))
 				{
 					throw DataException::forEmptyPrimaryKey('insertBatch');
-				}
-
-				// Set created_at and updated_at with same time
-				$date = $this->setDate();
-
-				if ($this->useTimestamps && $this->createdField && ! array_key_exists($this->createdField, $row))
-				{
-					$row[$this->createdField] = $date;
-				}
-
-				if ($this->useTimestamps && $this->updatedField && ! array_key_exists($this->updatedField, $row))
-				{
-					$row[$this->updatedField] = $date;
 				}
 			}
 		}
@@ -924,80 +338,14 @@ class Model
 	 * Updates a single record in $this->table. If an object is provided,
 	 * it will attempt to convert it into an array.
 	 *
-	 * @param integer|array|string|null $id
-	 * @param array|object|null         $data
+	 * @param integer|array|string|null $id     ID
+	 * @param array|object|null         $data   Data
+	 * @param boolean|null              $escape Escape
 	 *
 	 * @return boolean
-	 * @throws ReflectionException
 	 */
-	public function update($id = null, $data = null): bool
+	protected function _update($id = null, $data = null, ?bool $escape = null): bool
 	{
-		$escape = null;
-
-		if (is_numeric($id) || is_string($id))
-		{
-			$id = [$id];
-		}
-
-		if (empty($data))
-		{
-			$data           = $this->tempData['data'] ?? null;
-			$escape         = $this->tempData['escape'] ?? null;
-			$this->tempData = [];
-		}
-
-		if (empty($data))
-		{
-			throw DataException::forEmptyDataset('update');
-		}
-
-		// If $data is using a custom class with public or protected
-		// properties representing the table elements, we need to grab
-		// them as an array.
-		if (is_object($data) && ! $data instanceof stdClass)
-		{
-			$data = static::classToArray($data, $this->primaryKey, $this->dateFormat);
-		}
-
-		// If it's still a stdClass, go ahead and convert to
-		// an array so doProtectFields and other model methods
-		// don't have to do special checks.
-		if (is_object($data))
-		{
-			$data = (array) $data;
-		}
-
-		// If it's still empty here, means $data is no change or is empty object
-		if (empty($data))
-		{
-			throw DataException::forEmptyDataset('update');
-		}
-
-		// Validate data before saving.
-		if (! $this->skipValidation && ! $this->cleanRules(true)->validate($data))
-		{
-			return false;
-		}
-
-		// Must be called first so we don't
-		// strip out updated_at values.
-		$data = $this->doProtectFields($data);
-
-		if ($this->useTimestamps && $this->updatedField && ! array_key_exists($this->updatedField, $data))
-		{
-			$data[$this->updatedField] = $this->setDate();
-		}
-
-		$eventData = [
-			'id'   => $id,
-			'data' => $data,
-		];
-
-		if ($this->tempAllowCallbacks)
-		{
-			$eventData = $this->trigger('beforeUpdate', $eventData);
-		}
-
 		$builder = $this->builder();
 
 		if ($id)
@@ -1006,24 +354,9 @@ class Model
 		}
 
 		// Must use the set() method to ensure objects get converted to arrays
-		$result = $builder
-			->set($eventData['data'], '', $escape)
+		return $builder
+			->set($data, '', $escape)
 			->update();
-
-		$eventData = [
-			'id'     => $id,
-			'data'   => $eventData['data'],
-			'result' => $result,
-		];
-
-		if ($this->tempAllowCallbacks)
-		{
-			$this->trigger('afterUpdate', $eventData);
-		}
-
-		$this->tempAllowCallbacks = $this->allowCallbacks;
-
-		return $result;
 	}
 
 	/**
@@ -1031,62 +364,22 @@ class Model
 	 *
 	 * Compiles an update string and runs the query
 	 *
-	 * @param array   $set       An associative array of update values
-	 * @param string  $index     The where key
-	 * @param integer $batchSize The size of the batch to run
-	 * @param boolean $returnSQL True means SQL is returned, false will execute the query
+	 * @param array|null  $set       An associative array of update values
+	 * @param string|null $index     The where key
+	 * @param integer     $batchSize The size of the batch to run
+	 * @param boolean     $returnSQL True means SQL is returned, false will execute the query
 	 *
 	 * @return mixed    Number of rows affected or FALSE on failure
-	 * @throws DatabaseException
+	 * @throws DatabaseException DatabaseException.
+	 * @throws ReflectionException ReflectionException.
 	 */
-	public function updateBatch(array $set = null, string $index = null, int $batchSize = 100, bool $returnSQL = false)
+	protected function _updateBatch(
+		array $set = null,
+		string $index = null,
+		int $batchSize = 100,
+		bool $returnSQL = false
+	)
 	{
-		if (is_array($set))
-		{
-			foreach ($set as &$row)
-			{
-				// If $data is using a custom class with public or protected
-				// properties representing the table elements, we need to grab
-				// them as an array.
-				if (is_object($row) && ! $row instanceof stdClass)
-				{
-					$row = static::classToArray($row, $this->primaryKey, $this->dateFormat);
-				}
-
-				// If it's still a stdClass, go ahead and convert to
-				// an array so doProtectFields and other model methods
-				// don't have to do special checks.
-				if (is_object($row))
-				{
-					$row = (array) $row;
-				}
-
-				// Validate data before saving.
-				if (! $this->skipValidation && ! $this->cleanRules(true)->validate($row))
-				{
-					return false;
-				}
-
-				// Save updateIndex for later
-				$updateIndex = $row[$index] ?? null;
-
-				// Must be called first so we don't
-				// strip out updated_at values.
-				$row = $this->doProtectFields($row);
-
-				// Restore updateIndex value in case it was wiped out
-				if ($updateIndex !== null)
-				{
-					$row[$index] = $updateIndex;
-				}
-
-				if ($this->useTimestamps && $this->updatedField && ! array_key_exists($this->updatedField, $row))
-				{
-					$row[$this->updatedField] = $this->setDate();
-				}
-			}
-		}
-
 		return $this->builder()->testMode($returnSQL)->updateBatch($set, $index, $batchSize);
 	}
 
@@ -1098,30 +391,15 @@ class Model
 	 * @param boolean                   $purge Allows overriding the soft deletes setting.
 	 *
 	 * @return BaseResult|boolean
-	 * @throws DatabaseException
+	 * @throws DatabaseException DatabaseException.
 	 */
-	public function delete($id = null, bool $purge = false)
+	protected function _delete($id = null, bool $purge = false)
 	{
-		if ($id && (is_numeric($id) || is_string($id)))
-		{
-			$id = [$id];
-		}
-
 		$builder = $this->builder();
 
 		if ($id)
 		{
 			$builder = $builder->whereIn($this->primaryKey, $id);
-		}
-
-		$eventData = [
-			'id'    => $id,
-			'purge' => $purge,
-		];
-
-		if ($this->tempAllowCallbacks)
-		{
-			$this->trigger('beforeDelete', $eventData);
 		}
 
 		if ($this->useSoftDeletes && ! $purge)
@@ -1130,7 +408,9 @@ class Model
 			{
 				if (CI_DEBUG)
 				{
-					throw new DatabaseException('Deletes are not allowed unless they contain a "where" or "like" clause.');
+					throw new DatabaseException(
+						'Deletes are not allowed unless they contain a "where" or "like" clause.'
+					);
 				}
 
 				return false; // @codeCoverageIgnore
@@ -1150,20 +430,6 @@ class Model
 			$result = $builder->delete();
 		}
 
-		$eventData = [
-			'id'     => $id,
-			'purge'  => $purge,
-			'result' => $result,
-			'data'   => null,
-		];
-
-		if ($this->tempAllowCallbacks)
-		{
-			$this->trigger('afterDelete', $eventData);
-		}
-
-		$this->tempAllowCallbacks = $this->allowCallbacks;
-
 		return $result;
 	}
 
@@ -1173,45 +439,22 @@ class Model
 	 *
 	 * @return boolean|mixed
 	 */
-	public function purgeDeleted()
+	protected function _purgeDeleted()
 	{
-		if (! $this->useSoftDeletes)
-		{
-			return true;
-		}
-
 		return $this->builder()
 			->where($this->table . '.' . $this->deletedField . ' IS NOT NULL')
 			->delete();
 	}
 
 	/**
-	 * Sets $useSoftDeletes value so that we can temporarily override
-	 * the softdeletes settings. Can be used for all find* methods.
-	 *
-	 * @param boolean $val
-	 *
-	 * @return $this
-	 */
-	public function withDeleted(bool $val = true)
-	{
-		$this->tempUseSoftDeletes = ! $val;
-
-		return $this;
-	}
-
-	/**
 	 * Works with the find* methods to return only the rows that
 	 * have been deleted.
 	 *
-	 * @return $this
+	 * @return void
 	 */
-	public function onlyDeleted()
+	protected function _onlyDeleted()
 	{
-		$this->tempUseSoftDeletes = false;
 		$this->builder()->where($this->table . '.' . $this->deletedField . ' IS NOT NULL');
-
-		return $this;
 	}
 
 	/**
@@ -1219,19 +462,13 @@ class Model
 	 *
 	 * Compiles a replace into string and runs the query
 	 *
-	 * @param array|null $data
-	 * @param boolean    $returnSQL
+	 * @param array|null $data      Data
+	 * @param boolean    $returnSQL Set to true to return Query String
 	 *
 	 * @return mixed
 	 */
-	public function replace(array $data = null, bool $returnSQL = false)
+	protected function _replace(array $data = null, bool $returnSQL = false)
 	{
-		// Validate data before saving.
-		if ($data && ! $this->skipValidation && ! $this->cleanRules(true)->validate($data))
-		{
-			return false;
-		}
-
 		return $this->builder()->testMode($returnSQL)->replace($data);
 	}
 
@@ -1240,43 +477,16 @@ class Model
 	//--------------------------------------------------------------------
 
 	/**
-	 * Sets the return type of the results to be as an associative array.
-	 *
-	 * @return $this
-	 */
-	public function asArray()
-	{
-		$this->tempReturnType = 'array';
-
-		return $this;
-	}
-
-	/**
-	 * Sets the return type to be of the specified type of object.
-	 * Defaults to a simple object, but can be any class that has
-	 * class vars with the same name as the table columns, or at least
-	 * allows them to be created.
-	 *
-	 * @param string $class
-	 *
-	 * @return $this
-	 */
-	public function asObject(string $class = 'object')
-	{
-		$this->tempReturnType = $class;
-
-		return $this;
-	}
-
-	/**
 	 * Loops over records in batches, allowing you to operate on them.
 	 * Works with $this->builder to get the Compiled select to
 	 * determine the rows to operate on.
 	 *
-	 * @param integer $size
-	 * @param Closure $userFunc
+	 * @param integer $size     Size
+	 * @param Closure $userFunc Callback Function
 	 *
-	 * @throws DataException
+	 * @throws DataException DataException.
+	 *
+	 * @return void
 	 */
 	public function chunk(int $size, Closure $userFunc)
 	{
@@ -1310,54 +520,6 @@ class Model
 				}
 			}
 		}
-	}
-
-	/**
-	 * Works with $this->builder to get the Compiled Select to operate on.
-	 * Expects a GET variable (?page=2) that specifies the page of results
-	 * to display.
-	 *
-	 * @param integer|null $perPage
-	 * @param string       $group   Will be used by the pagination library to identify a unique pagination set.
-	 * @param integer|null $page    Optional page number (useful when the page number is provided in different way)
-	 * @param integer      $segment Optional URI segment number (if page number is provided by URI segment)
-	 *
-	 * @return array|null
-	 */
-	public function paginate(int $perPage = null, string $group = 'default', int $page = null, int $segment = 0)
-	{
-		$pager = Services::pager(null, null, false);
-
-		if ($segment)
-		{
-			$pager->setSegment($segment);
-		}
-
-		$page  = $page >= 1 ? $page : $pager->getCurrentPage($group);
-		$total = $this->countAllResults(false);
-
-		// Store it in the Pager library so it can be
-		// paginated in the views.
-		$this->pager = $pager->store($group, $page, $perPage, $total, $segment);
-		$perPage     = $this->pager->getPerPage($group);
-		$offset      = ($page - 1) * $perPage;
-
-		return $this->findAll($perPage, $offset);
-	}
-
-	/**
-	 * Sets whether or not we should whitelist data set during
-	 * updates or inserts against $this->availableFields.
-	 *
-	 * @param boolean $protect
-	 *
-	 * @return $this
-	 */
-	public function protect(bool $protect = true)
-	{
-		$this->protectFields = $protect;
-
-		return $this;
 	}
 
 	/**
@@ -1410,397 +572,20 @@ class Model
 	}
 
 	/**
-	 * Ensures that only the fields that are allowed to be updated
-	 * are in the data array.
-	 *
-	 * Used by insert() and update() to protect against mass assignment
-	 * vulnerabilities.
-	 *
-	 * @param array $data
-	 *
-	 * @return array
-	 * @throws DataException
-	 */
-	protected function doProtectFields(array $data): array
-	{
-		if (! $this->protectFields)
-		{
-			return $data;
-		}
-
-		if (empty($this->allowedFields))
-		{
-			throw DataException::forInvalidAllowedFields(get_class($this));
-		}
-
-		foreach ($data as $key => $val)
-		{
-			if (! in_array($key, $this->allowedFields, true))
-			{
-				unset($data[$key]);
-			}
-		}
-
-		return $data;
-	}
-
-	/**
-	 * A utility function to allow child models to use the type of
-	 * date/time format that they prefer. This is primarily used for
-	 * setting created_at, updated_at and deleted_at values, but can be
-	 * used by inheriting classes.
-	 *
-	 * The available time formats are:
-	 *  - 'int'      - Stores the date as an integer timestamp
-	 *  - 'datetime' - Stores the data in the SQL datetime format
-	 *  - 'date'     - Stores the date (only) in the SQL date format.
-	 *
-	 * @param integer $userData An optional PHP timestamp to be converted.
-	 *
-	 * @return mixed
-	 * @throws ModelException ;
-	 */
-	protected function setDate(int $userData = null)
-	{
-		$currentDate = $userData ?? time();
-
-		switch ($this->dateFormat)
-		{
-			case 'int':
-				return $currentDate;
-			case 'datetime':
-				return date('Y-m-d H:i:s', $currentDate);
-			case 'date':
-				return date('Y-m-d', $currentDate);
-			default:
-				throw ModelException::forNoDateFormat(static::class);
-		}
-	}
-
-	/**
-	 * Specify the table associated with a model
-	 *
-	 * @param string $table
-	 *
-	 * @return $this
-	 */
-	public function setTable(string $table)
-	{
-		$this->table = $table;
-
-		return $this;
-	}
-
-	/**
-	 * Grabs the last error(s) that occurred. If data was validated,
-	 * it will first check for errors there, otherwise will try to
-	 * grab the last error from the Database connection.
-	 *
-	 * @param boolean $forceDB Always grab the db error, not validation
+	 * Grabs the last error(s) that occurred from the Database connection.
 	 *
 	 * @return array|null
 	 */
-	public function errors(bool $forceDB = false)
+	protected function _errors()
 	{
-		// Do we have validation errors?
-		if (! $forceDB && ! $this->skipValidation)
-		{
-			$errors = $this->validation->getErrors();
-
-			if (! empty($errors))
-			{
-				return $errors;
-			}
-		}
-
-		$error = $this->db->error();
-
-		return $error['message'] ?? null;
-	}
-
-	/**
-	 * It could be used when you have to change default or override current allowed fields.
-	 *
-	 * @param array $allowedFields
-	 *
-	 * @return $this
-	 */
-	public function setAllowedFields(array $allowedFields)
-	{
-		$this->allowedFields = $allowedFields;
-
-		return $this;
-	}
-
-	//--------------------------------------------------------------------
-	// Validation
-	//--------------------------------------------------------------------
-
-	/**
-	 * Set the value of the skipValidation flag.
-	 *
-	 * @param boolean $skip
-	 *
-	 * @return $this
-	 */
-	public function skipValidation(bool $skip = true)
-	{
-		$this->skipValidation = $skip;
-
-		return $this;
-	}
-
-	/**
-	 * Allows to set validation messages.
-	 * It could be used when you have to change default or override current validate messages.
-	 *
-	 * @param array $validationMessages
-	 *
-	 * @return $this
-	 */
-	public function setValidationMessages(array $validationMessages)
-	{
-		$this->validationMessages = $validationMessages;
-
-		return $this;
-	}
-
-	/**
-	 * Allows to set field wise validation message.
-	 * It could be used when you have to change default or override current validate messages.
-	 *
-	 * @param string $field
-	 * @param array  $fieldMessages
-	 *
-	 * @return $this
-	 */
-	public function setValidationMessage(string $field, array $fieldMessages)
-	{
-		$this->validationMessages[$field] = $fieldMessages;
-
-		return $this;
-	}
-
-	/**
-	 * Allows to set validation rules.
-	 * It could be used when you have to change default or override current validate rules.
-	 *
-	 * @param array $validationRules
-	 *
-	 * @return $this
-	 */
-	public function setValidationRules(array $validationRules)
-	{
-		$this->validationRules = $validationRules;
-
-		return $this;
-	}
-
-	/**
-	 * Allows to set field wise validation rules.
-	 * It could be used when you have to change default or override current validate rules.
-	 *
-	 * @param string       $field
-	 * @param string|array $fieldRules
-	 *
-	 * @return $this
-	 */
-	public function setValidationRule(string $field, $fieldRules)
-	{
-		$this->validationRules[$field] = $fieldRules;
-
-		return $this;
-	}
-
-	/**
-	 * Should validation rules be removed before saving?
-	 * Most handy when doing updates.
-	 *
-	 * @param boolean $choice
-	 *
-	 * @return $this
-	 */
-	public function cleanRules(bool $choice = false)
-	{
-		$this->cleanValidationRules = $choice;
-
-		return $this;
-	}
-
-	/**
-	 * Validate the data against the validation rules (or the validation group)
-	 * specified in the class property, $validationRules.
-	 *
-	 * @param array|object $data
-	 *
-	 * @return boolean
-	 */
-	public function validate($data): bool
-	{
-		$rules = $this->getValidationRules();
-
-		if ($this->skipValidation || empty($rules) || empty($data))
-		{
-			return true;
-		}
-
-		// Query Builder works with objects as well as arrays,
-		// but validation requires array, so cast away.
-		if (is_object($data))
-		{
-			$data = (array) $data;
-		}
-
-		$rules = $this->cleanValidationRules ? $this->cleanValidationRules($rules, $data) : $rules;
-
-		// If no data existed that needs validation
-		// our job is done here.
-		if (empty($rules))
-		{
-			return true;
-		}
-
-		return $this->validation
-			->setRules($rules, $this->validationMessages)
-			->run($data, null, $this->DBGroup);
-	}
-
-	/**
-	 * Removes any rules that apply to fields that have not been set
-	 * currently so that rules don't block updating when only updating
-	 * a partial row.
-	 *
-	 * @param array      $rules
-	 * @param array|null $data
-	 *
-	 * @return array
-	 */
-	protected function cleanValidationRules(array $rules, array $data = null): array
-	{
-		if (empty($data))
-		{
-			return [];
-		}
-
-		foreach ($rules as $field => $rule)
-		{
-			if (! array_key_exists($field, $data))
-			{
-				unset($rules[$field]);
-			}
-		}
-
-		return $rules;
-	}
-
-	/**
-	 * Replace any placeholders within the rules with the values that
-	 * match the 'key' of any properties being set. For example, if
-	 * we had the following $data array:
-	 *
-	 * [ 'id' => 13 ]
-	 *
-	 * and the following rule:
-	 *
-	 *  'required|is_unique[users,email,id,{id}]'
-	 *
-	 * The value of {id} would be replaced with the actual id in the form data:
-	 *
-	 *  'required|is_unique[users,email,id,13]'
-	 *
-	 * @codeCoverageIgnore
-	 *
-	 * @deprecated use fillPlaceholders($rules, $data) from Validation instead
-	 *
-	 * @param array $rules
-	 * @param array $data
-	 *
-	 * @return array
-	 */
-	protected function fillPlaceholders(array $rules, array $data): array
-	{
-		$replacements = [];
-
-		foreach ($data as $key => $value)
-		{
-			$replacements["{{$key}}"] = $value;
-		}
-
-		if (! empty($replacements))
-		{
-			foreach ($rules as &$rule)
-			{
-				if (is_array($rule))
-				{
-					foreach ($rule as &$row)
-					{
-						// Should only be an `errors` array
-						// which doesn't take placeholders.
-						if (is_array($row))
-						{
-							continue;
-						}
-
-						$row = strtr($row, $replacements);
-					}
-
-					continue;
-				}
-
-				$rule = strtr($rule, $replacements);
-			}
-		}
-
-		return $rules;
-	}
-
-	/**
-	 * Returns the model's defined validation rules so that they
-	 * can be used elsewhere, if needed.
-	 *
-	 * @param array $options
-	 *
-	 * @return array
-	 */
-	public function getValidationRules(array $options = []): array
-	{
-		$rules = $this->validationRules;
-
-		// ValidationRules can be either a string, which is the group name,
-		// or an array of rules.
-		if (is_string($rules))
-		{
-			$rules = $this->validation->loadRuleGroup($rules); // @phpstan-ignore-line
-		}
-
-		if (isset($options['except']))
-		{
-			$rules = array_diff_key($rules, array_flip($options['except']));
-		}
-		elseif (isset($options['only']))
-		{
-			$rules = array_intersect_key($rules, array_flip($options['only']));
-		}
-
-		return $rules;
-	}
-
-	/**
-	 * Returns the model's define validation messages so they
-	 * can be used elsewhere, if needed.
-	 *
-	 * @return array
-	 */
-	public function getValidationMessages(): array
-	{
-		return $this->validationMessages;
+		return $this->db->error();
 	}
 
 	/**
 	 * Override countAllResults to account for soft deleted accounts.
 	 *
-	 * @param boolean $reset
-	 * @param boolean $test
+	 * @param boolean $reset Reset
+	 * @param boolean $test  Test
 	 *
 	 * @return mixed
 	 */
@@ -1821,63 +606,6 @@ class Model
 		return $this->builder()->testMode($test)->countAllResults($reset);
 	}
 
-	/**
-	 * Sets $tempAllowCallbacks value so that we can temporarily override
-	 * the setting. Resets after the next method that uses triggers.
-	 *
-	 * @param boolean $val
-	 *
-	 * @return $this
-	 */
-	public function allowCallbacks(bool $val = true)
-	{
-		$this->tempAllowCallbacks = $val;
-
-		return $this;
-	}
-
-	/**
-	 * A simple event trigger for Model Events that allows additional
-	 * data manipulation within the model. Specifically intended for
-	 * usage by child models this can be used to format data,
-	 * save/load related classes, etc.
-	 *
-	 * It is the responsibility of the callback methods to return
-	 * the data itself.
-	 *
-	 * Each $eventData array MUST have a 'data' key with the relevant
-	 * data for callback methods (like an array of key/value pairs to insert
-	 * or update, an array of results, etc)
-	 *
-	 * If callbacks are not allowed then returns $eventData immediately.
-	 *
-	 * @param string $event
-	 * @param array  $eventData
-	 *
-	 * @return mixed
-	 * @throws DataException
-	 */
-	protected function trigger(string $event, array $eventData)
-	{
-		// Ensure it's a valid event
-		if (! isset($this->{$event}) || empty($this->{$event}))
-		{
-			return $eventData;
-		}
-
-		foreach ($this->{$event} as $callback)
-		{
-			if (! method_exists($this, $callback))
-			{
-				throw DataException::forInvalidMethodTriggered($callback);
-			}
-
-			$eventData = $this->{$callback}($eventData);
-		}
-
-		return $eventData;
-	}
-
 	//--------------------------------------------------------------------
 	// Magic
 	//--------------------------------------------------------------------
@@ -1885,20 +613,15 @@ class Model
 	/**
 	 * Provides/instantiates the builder/db connection and model's table/primary key names and return type.
 	 *
-	 * @param string $name
+	 * @param string $name Name
 	 *
 	 * @return mixed
 	 */
 	public function __get(string $name)
 	{
-		if (property_exists($this, $name))
+		if (parent::__isset($name))
 		{
-			return $this->$name;
-		}
-
-		if (isset($this->db->$name))
-		{
-			return $this->db->$name;
+			return parent::__get($name);
 		}
 
 		if (isset($this->builder()->$name))
@@ -1912,18 +635,13 @@ class Model
 	/**
 	 * Checks for the existence of properties across this model, builder, and db connection.
 	 *
-	 * @param string $name
+	 * @param string $name Name
 	 *
 	 * @return boolean
 	 */
 	public function __isset(string $name): bool
 	{
-		if (property_exists($this, $name))
-		{
-			return true;
-		}
-
-		if (isset($this->db->$name))
+		if (parent::__isset($name))
 		{
 			return true;
 		}
@@ -1940,20 +658,16 @@ class Model
 	 * Provides direct access to method in the builder (if available)
 	 * and the database connection.
 	 *
-	 * @param string $name
-	 * @param array  $params
+	 * @param string $name   Name
+	 * @param array  $params Params
 	 *
 	 * @return $this|null
 	 */
 	public function __call(string $name, array $params)
 	{
-		$result = null;
+		$result = parent::__call($name, $params);
 
-		if (method_exists($this->db, $name))
-		{
-			$result = $this->db->{$name}(...$params);
-		}
-		elseif (method_exists($builder = $this->builder(), $name))
+		if ($result === null && method_exists($builder = $this->builder(), $name))
 		{
 			$result = $builder->{$name}(...$params);
 		}
@@ -1964,17 +678,17 @@ class Model
 			{
 				$className = static::class;
 
-				throw new BadMethodCallException("Call to undefined method $className::$name");
+				throw new BadMethodCallException('Call to undefined method ' . $className . '::' . $name);
 			}
 
 			return $result;
 		}
 
-		if (! $result instanceof BaseBuilder)
+		if ($result instanceof BaseBuilder)
 		{
-			return $result;
+			return $this;
 		}
 
-		return $this;
+		return $result;
 	}
 }
